services/maya-swarm/swarm_manager.py
from camel.agents import ChatAgent
from camel.messages import BaseMessage
from camel.models import ModelFactory
from camel.types import ModelType
import os

class SwarmManager:
    def __init__(self):
        self.role_configs = {
            "coder": "Expert Software Engineer. Specializes in Python, Rust, and TypeScript.",
            "researcher": "Expert Analyst. Specializes in deep-web search and data synthesis.",
            "architect": "System Architect. Specializes in high-level design and swarm orchestration."
        }
        
    async def execute_task(self, prompt, byok_key=None):
        # 1. Task Decomposition (The OWL Logic)
        # For prototype: simple routing based on keywords
        target_role = "architect"
        if "code" in prompt.lower() or "build" in prompt.lower():
            target_role = "coder"
        elif "search" in prompt.lower() or "find" in prompt.lower():
            target_role = "researcher"

        # 2. Agent Initialization
        role_desc = self.role_configs.get(target_role)
        
        # 3. Execution
        return f"MAYA Swarm [{target_role.upper()}]: Analyzing task -> '{prompt}'. Using specialized persona: {role_desc}"
