services/maya-swarm/requirements.txt
fastapi
uvicorn
pydantic
requests
camel-ai[all]
mem0ai
supabase
python-dotenv
